refactor: remove time travel experiment as it is now stable
