+++ lib/constants/local-storage.ts
@@ -2,8 +2,6 @@ export enum LocalStorageKey {
     UIMode = 'ui-mode'
 }
 
-export const getTimeTravelFeatureLocalStorageKey = (toolName: string): string => `time-travel-${toolName}`;
-
 export const TIME_TRAVEL_PLAYER_VISIBILITY_KEY = 'time-travel-player-visibility';
 
 export enum UiMode {
