+++ lib/static/modules/reducers/features.ts
@@ -2,9 +2,7 @@ import {State} from '@/static/new-ui/types/store';
 import {SomeAction} from '@/static/modules/actions/types';
 import actionNames from '@/static/modules/action-names';
 import {applyStateUpdate} from '@/static/modules/utils';
-import {Feature, ShowTimeTravelExperimentFeature, TimeTravelFeature, ToolName} from '@/constants';
-import * as localStorageWrapper from '../local-storage-wrapper';
-import {getTimeTravelFeatureLocalStorageKey} from '@/constants/local-storage';
+import {Feature, TimeTravelFeature, ToolName} from '@/constants';
 
 export default (state: State, action: SomeAction): State => {
     switch (action.type) {
@@ -14,12 +12,7 @@ export default (state: State, action: SomeAction): State => {
 
             const toolName = action.payload.apiValues?.toolName;
             if (toolName === ToolName.Testplane) {
-                features.push(ShowTimeTravelExperimentFeature);
-
-                const isTimeTravelEnabled = localStorageWrapper.getItem(getTimeTravelFeatureLocalStorageKey(toolName ?? ''), true);
-                if (isTimeTravelEnabled) {
-                    features.push(TimeTravelFeature);
-                }
+                features.push(TimeTravelFeature);
             }
 
             if (features.length > 0) {
@@ -29,9 +22,6 @@ export default (state: State, action: SomeAction): State => {
             return state;
         }
         case actionNames.SET_AVAILABLE_FEATURES: {
-            const toolName = state.apiValues.toolName;
-            localStorageWrapper.setItem(getTimeTravelFeatureLocalStorageKey(toolName), action.payload.features.some(f => f.name === TimeTravelFeature.name));
-
             return applyStateUpdate(state, {app: {availableFeatures: action.payload.features}});
         }
     }
