+++ lib/static/new-ui/features/suites/components/SnapshotsPlayer/PlayIcon.tsx
