+++ lib/static/new-ui/features/suites/components/SnapshotsPlayer/Timeline.tsx
