+++ lib/static/new-ui/features/suites/components/SnapshotsPlayer/index.module.css
