+++ lib/static/new-ui/features/suites/components/SnapshotsPlayer/types.ts
