+++ lib/static/new-ui/features/suites/components/SnapshotsPlayer/utils.ts
