+++ lib/static/new-ui/features/suites/components/TestInfo/index.tsx
@@ -1,16 +1,17 @@
+import {Spin} from '@gravity-ui/uikit';
+import classNames from 'classnames';
 import React, {ReactNode} from 'react';
+import {useSelector} from 'react-redux';
+
 import {CollapsibleSection} from '@/static/new-ui/features/suites/components/CollapsibleSection';
 import {MetaInfo} from '@/static/new-ui/components/MetaInfo';
 import {TestSteps} from '@/static/new-ui/features/suites/components/TestSteps';
 import {getCurrentResult, isTimeTravelPlayerAvailable} from '@/static/new-ui/features/suites/selectors';
-import {useSelector} from 'react-redux';
+import {getTestSteps} from '@/static/new-ui/features/suites/components/TestSteps/selectors';
+import {SnapshotsPlayer} from '@/static/new-ui/features/suites/components/SnapshotsPlayer';
+import {ErrorHandler} from '@/static/new-ui/features/error-handling/components/ErrorHandling';
 
 import styles from './index.module.css';
-import {Spin} from '@gravity-ui/uikit';
-
-import {getTestSteps} from '@/static/new-ui/features/suites/components/TestSteps/selectors';
-import {SnapshotsPlayer} from '@/static/new-ui/experiments/time-travel/components/SnapshotsPlayer';
-import classNames from 'classnames';
 
 export function TestInfo(): ReactNode {
     const currentResult = useSelector(getCurrentResult);
@@ -26,7 +27,9 @@ export function TestInfo(): ReactNode {
         <CollapsibleSection id={'actions'} title={'Actions'}>
             <div className={styles.stepsContainer}>
                 {steps.length > 0 ?
-                    <TestSteps className={styles.stepsListContainer}/> :
+                    <ErrorHandler.Boundary watchFor={[currentResult]} fallback={<ErrorHandler.FallbackDataCorruption />}>
+                        <TestSteps className={styles.stepsListContainer}/>
+                    </ErrorHandler.Boundary> :
                     <div className={styles.emptyStepsContainer}>{isRunning ? <><Spin size={'xs'} style={{marginRight: '4px'}} />Test is running</> : 'No steps to show'}</div>
                 }
                 {<div className={classNames(styles.sticky, !shouldShowPlayer && styles.hidden)}>
